refactor: move permissions decorator and guard into permissions directory
